CIF-1042 - Add PaymentProvider component and form validation
