+++ ui.apps/src/main/javascript/minicart/src/components/Checkout/paymentProviders/paymentProvider.css
@@ -1,0 +1,6 @@
+.braintree {
+    composes: braintree from '../paymentsForm.css';
+}
+.error_message {
+    composes: root_error from '../../Field/message.css';
+}
